Moved the methods exctracted from the view to util
