+++ corehq/apps/importer/views.py
@@ -1,22 +1,20 @@
 import os.path
 from django.http import HttpResponseRedirect
-from casexml.apps.case.models import CommCareCase, const
+from casexml.apps.case.models import CommCareCase
 from casexml.apps.phone.xml import get_case_xml
 from corehq.apps.hqcase.utils import submit_case_blocks
 from corehq.apps.importer import base
 from corehq.apps.importer.const import LookupErrors
-from corehq.apps.importer.util import ExcelFile, get_case_properties
+from corehq.apps.importer.util import *
 from couchdbkit.exceptions import MultipleResultsFound, NoResultFound
 from django.views.decorators.http import require_POST
 from datetime import datetime, date
-from dimagi.utils.parsing import json_format_datetime
 from xlrd import xldate_as_tuple
 from corehq.apps.users.decorators import require_permission
 from corehq.apps.users.models import Permissions
 from soil.util import expose_download
 from soil import DownloadBase
 
-from xml.etree import ElementTree
 from casexml.apps.case.tests.util import CaseBlock
 import uuid
 
@@ -169,120 +167,6 @@ def excel_fields(request, domain):
                                     'name': 'Import: Match columns to fields'
                                  },
                                 'slug': base.ImportCases.slug})
-
-def convert_custom_fields_to_struct(request):
-    # TODO musn't be able to select an excel_field twice (in html)
-    excel_fields = request.POST.getlist('excel_field[]')
-    case_fields = request.POST.getlist('case_field[]')
-    custom_fields = request.POST.getlist('custom_field[]')
-    date_yesno = request.POST.getlist('date_yesno[]')
-
-    field_map = {}
-    for i, field in enumerate(excel_fields):
-        if field and (case_fields[i] or custom_fields[i]):
-            field_map[field] = {'case': case_fields[i],
-                                'custom': custom_fields[i],
-                                'date': int(date_yesno[i])}
-
-    return field_map
-
-def parse_excel_date(date):
-    return date(*xldate_as_tuple(date, 0)[:3])
-
-def parse_search_id(request, columns, row):
-    # Find index of user specified search column
-    search_column = request.POST['search_column']
-    search_column_index = columns.index(search_column)
-
-    search_id = row[search_column_index]
-    try:
-        # if the spreadsheet gives a number, strip any decimals off
-        # float(x) is more lenient in conversion from string so both
-        # are used
-        search_id = int(float(search_id))
-    except ValueError:
-        # if it's not a number that's okay too
-        pass
-
-    # need a string no matter what the type was
-    return str(search_id)
-
-def submit_case_block(caseblock, domain, username, user_id):
-    casexml = ElementTree.tostring(caseblock.as_xml(format_datetime=json_format_datetime))
-    submit_case_blocks(casexml, domain, username, user_id)
-
-def get_key_column_index(request, columns):
-    key_column = request.POST['key_column']
-    try:
-        key_column_index = columns.index(key_column)
-    except ValueError:
-        key_column_index = False
-
-    return key_column_index
-
-def get_value_column_index(request, columns):
-    value_column = request.POST['value_column']
-    try:
-        value_column_index = columns.index(value_column)
-    except ValueError:
-        value_column_index = False
-
-    return value_column_index
-
-def lookup_case(search_field, search_id, domain):
-    found = False
-    if search_field == 'case_id':
-        try:
-            case = CommCareCase.get(search_id)
-            if case.domain == domain:
-                found = True
-        except Exception:
-            pass
-    elif search_field == 'external_id':
-        try:
-            case = CommCareCase.view('hqcase/by_domain_external_id',
-                                     key=[domain, search_id],
-                                     reduce=False,
-                                     include_docs=True).one()
-            found = bool(case)
-        except NoResultFound:
-            pass
-        except MultipleResultsFound:
-            return (None, LookupErrors.MultipleResults)
-
-    if found:
-        return (case, None)
-    else:
-        return (None, LookupErrors.NotFound)
-
-def populate_updated_fields(request, columns, row):
-    field_map = convert_custom_fields_to_struct(request)
-    key_column_index = get_key_column_index(request, columns)
-    value_column_index = get_value_column_index(request, columns)
-    fields_to_update = {}
-
-    for key in field_map:
-        try:
-            if key_column_index and key == row[key_column_index]:
-                update_value = row[value_column_index]
-            else:
-                update_value = row[columns.index(key)]
-        except:
-            continue
-
-        if field_map[key]['custom']:
-            # custom (new) field was entered
-            update_field_name = field_map[key]['custom']
-        else:
-            # existing case field was chosen
-            update_field_name = field_map[key]['case']
-
-        if field_map[key]['date'] == 1:
-            update_value = parse_excel_date(update_value)
-
-        fields_to_update[update_field_name] = update_value
-
-    return fields_to_update
 
 @require_POST
 @require_can_edit_data
